requirements.txt
# Core Python dependencies
openai>=1.3.0
python-dotenv>=1.0.0

# Google/YouTube API dependencies
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.86.0
httplib2>=0.22.0

# Speech-to-Text dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
faster-whisper>=0.10.0
openai-whisper>=20231117

# Audio/Video processing
moviepy>=1.0.3
pydub>=0.25.1
scipy>=1.11.0
numpy>=1.24.0
av>=10.0.0

# Diarization (speaker identification)
pyannote.audio>=3.0.0
speechbrain>=0.5.16

# Progress bars and utilities
tqdm>=4.65.0
requests>=2.28.0

# Additional audio libraries
librosa>=0.10.0
soundfile>=0.12.0

# For Windows users - audio dependencies
# pyaudio>=0.2.11  # Uncomment if needed, but can be tricky on Windows

# Optional: GPU acceleration (comment out if using CPU only)
# nvidia-cudnn-cu11>=8.6.0.163
# nvidia-cublas-cu11>=11.10.3.66